requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
torch>=2.0.0
transformers>=4.21.0
requests>=2.31.0
plotly>=5.15.0
statsmodels>=0.14.0
scipy>=1.10.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-decouple>=3.8
git+https://github.com/IBM/tsfm.git

# Production server
gunicorn>=21.2.0

# System monitoring
psutil>=5.9.0